alpha (13.2) fix - September 17, 2026 - updated the itemTable parameter type in makeForageSpot to use double instead of int - deleted the values to percentages conversion in makeForageSpot - updated comment in ForageSpot component to reflect new use for itemTable

+++ src/Game/Headers/Components.hpp
@@ -134,9 +134,8 @@ namespace Component
 	struct ForageSpot
 	{
 		double forageDistance = 0.0;
-		// key: item type, value: probabily percentage
-		// all doubles must add up to 1.0
-		std::unordered_map<Enum::Item, double> itemTable{};
+		// key: item type, value: probability weight
+		std::unordered_map<Enum::Item, int> itemTable{};
 	};
 
 	struct Delete
